split stuff up a bit
